refactor to use an extension for GLMakie
